refactor : 패키지 구조 리팩토링 - adapter, application 내부 도메인 패키지 생성
