+++ booquest-api/src/test/java/com/booquest/booquest_api/auth/application/service/SocialLoginServiceTest.java
@@ -4,12 +4,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.booquest.booquest_api.adapter.in.web.dto.SocialLoginResponse;
-import com.booquest.booquest_api.application.port.in.SocialLoginUseCase;
-import com.booquest.booquest_api.application.port.out.JwtTokenPort;
-import com.booquest.booquest_api.application.port.out.OAuthClientPort;
-import com.booquest.booquest_api.application.port.out.UserQueryPort;
-import com.booquest.booquest_api.application.service.SocialLoginService;
+import com.booquest.booquest_api.adapter.in.auth.web.dto.SocialLoginResponse;
+import com.booquest.booquest_api.application.port.in.auth.SocialLoginUseCase;
+import com.booquest.booquest_api.application.port.out.auth.JwtTokenPort;
+import com.booquest.booquest_api.application.port.out.auth.OAuthClientPort;
+import com.booquest.booquest_api.application.port.out.auth.UserQueryPort;
+import com.booquest.booquest_api.application.service.auth.SocialLoginService;
 import com.booquest.booquest_api.domain.user.model.SocialUser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
